Separated movies and auth into two services
