+++ auth/src/auth.js
@@ -41,12 +41,7 @@ const authFactory = (secret) => (username, password) => {
   );
 };
 
-const verifyUserFactory = (secret) => (token) => {
-  return jwt.verify(token, secret);
-};
-
 export {
   authFactory,
-  verifyUserFactory,
   AuthError
 };
