+++ movies/Dockerfile
@@ -1,0 +1,11 @@
+FROM node:14.15-alpine
+
+WORKDIR /app
+
+COPY ./package.json ./package-lock.json ./
+RUN npm install
+
+RUN mkdir ./src
+COPY ./src ./src
+
+CMD ["node", "./src/server.js"]
