+++ movies/docker-compose.yml
@@ -6,9 +6,10 @@ services:
       context: .
       dockerfile: Dockerfile
     ports:
-      - published: ${APP_PORT:-3000}
-        target: 3000
+      - published: ${APP_PORT:-3001}
+        target: 3001
     volumes:
       - .:/app/
     environment:
       JWT_SECRET: ${JWT_SECRET}
+      OMDB_API_KEY: ${OMDB_API_KEY}
