+++ movies/package.json
@@ -11,6 +11,7 @@
   "license": "",
   "dependencies": {
     "express": "4.17.1",
-    "jsonwebtoken": "^8.5.1"
+    "jsonwebtoken": "^8.5.1",
+    "node-fetch": "^2.6.1"
   }
 }
