+++ movies/src/auth.js
@@ -1,0 +1,7 @@
+import jwt from "jsonwebtoken";
+
+const verifyUserFactory = (secret) => (token) => {
+  return jwt.verify(token, secret);
+};
+
+export default verifyUserFactory;
